feat: improve scoring consistency, add decimal scores, and enhance loading UX
Reduce score variance across runs by fixing extraction at 20 requirements
with deterministic category weights and stricter match-level definitions.
Switch fit/confidence scores from int to float for accuracy. Replace LLM
confidence call with deterministic heuristics to reduce generation time.
Add progress bar with elapsed timer and time estimate to loading state.

+++ backend/app/models/report_schema.py
@@ -26,8 +26,8 @@ class ExecutiveSummary(BaseModel):
 
 
 class PursuitRecommendation(BaseModel):
-    overall_fit_score: int = Field(ge=0, le=10)
-    confidence_score: int = Field(ge=0, le=10)
+    overall_fit_score: float = Field(ge=0, le=10)
+    confidence_score: float = Field(ge=0, le=10)
     recommendation: str  # pursue | pursue_selectively | low_priority | insufficient_information
     reasoning: list[str]
     ideal_candidate_profile: str
